+++ source/Public/Install-GuestConfigurationAgent.ps1
@@ -1,7 +1,6 @@
 function Install-GuestConfigurationAgent
 {
     [CmdletBinding()]
-    [Experimental('GuestConfiguration.SetScenario', 'Show')]
     [OutputType([void])]
     param
     (