+++ source/Public/New-GuestConfigurationPackage.ps1
@@ -68,7 +68,6 @@ function New-GuestConfigurationPackage
         $Path = '.',
 
         [Parameter()]
-        [Experimental('GuestConfiguration.SetScenario', 'Show')]
         [PackageType]
         $Type = 'Audit',